Combined Core + Robot + Fauxbot
Use a combined system so that fauxbot "emulation" can be used on the
same codebase as we are using for the robot without requiring any
copy/paste between repositories.  Providing this emulation here instead
of in a separate branch will help to make exercises feel more "real", as
they really could be deployed onto a robot by running just a slightly
different gradle task.